requirements.txt
matplotlib
numpy
wandb
pillow
albumentations
pandas
requests
opencv-python
tqdm
scikit-image

# For efficient zarr loading
tifffile
zarr==2.16.1
imagecodecs
numcodecs==0.12.1
dask

vesuvius-phalanx
torchmetrics
scikit-learn
transformers
einops
lightning
lightning-fabric

h5py
batchgenerators
monai
beautifulsoup4
termcolor
future
timm
segmentation_models_pytorch
argparse
scipy
imageio

# for nnUNet
nnunetv2
dotenv
rich